+++ tests/func/test_checkout.py
@@ -85,7 +85,7 @@ class TestCheckoutCorruptedCacheFile(TestRepro):
 
 class TestCheckoutCorruptedCacheDir(TestDvc):
     def test(self):
-        from dvc.objects import load
+        from dvc.data import load
 
         # NOTE: using 'copy' so that cache and link don't have same inode
         ret = main(["config", "cache.type", "copy"])
