+++ tests/func/test_repo.py
@@ -1,5 +1,5 @@
+from dvc.data.db import ODBManager
 from dvc.dvcfile import PIPELINE_FILE, PIPELINE_LOCK
-from dvc.objects.db import ODBManager
 from dvc.system import System
