+++ tests/unit/fs/test_repo.py
@@ -4,9 +4,9 @@ from unittest import mock
 
 import pytest
 
+from dvc.data.stage import stage
 from dvc.fs.repo import RepoFileSystem
 from dvc.hash_info import HashInfo
-from dvc.objects.stage import stage
 from tests.utils import clean_staging
 
 
@@ -584,9 +584,9 @@ def test_get_hash_mixed_dir(tmp_dir, scm, dvc):
 
 
 def test_get_hash_dirty_file(tmp_dir, dvc):
-    from dvc.objects import check
+    from dvc.data import check
+    from dvc.data.stage import get_file_hash
     from dvc.objects.errors import ObjectFormatError
-    from dvc.objects.stage import get_file_hash
 
     tmp_dir.dvc_gen("file", "file")
     file_hash_info = HashInfo("md5", "8c7dd922ad47494fc02c388e12c00eac")
