+++ tests/unit/remote/test_remote.py
@@ -2,10 +2,10 @@ import os
 
 import pytest
 
+from dvc.data.db import get_index
 from dvc.fs import get_cloud_fs
 from dvc.fs.gs import GSFileSystem
 from dvc.fs.s3 import S3FileSystem
-from dvc.objects.db import get_index
 
 
 def test_remote_with_hash_jobs(dvc):
